backend/requirements.txt
django==4.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt